Enhanced MinHeap and Dijsktra's algorithm.
MinHeap had a slight bug fix - now indices corresponding to nodes that
haven't been added to the heap will correctly display as not in the
heap. Also, moved the movement cost to a cell calculation in Dijkstra's
algorithm to a separate function where it will be easier to make
changes, like accounting for other creatures present in the grid cell.