Added outlining possible moves for square grids
Now, the SquareMoveRenderer class can be used to outline all of the
possible moves a creature can make. I still have to make a decision on
wether or not to outline "islands" of impassable squares within the
outline - currently it does not do this. Also, there is some messy test
code that should be changed.