Added SQLite support
Added simple SQLite support, which is unfinished. This is just the proof
that we can add a database to the project and write to / read from the
database at runtime. More work needs to be done to encapsulate the
methods for accessing the database, as well as adding functionality to
the database itself.